勉强使用

+++ lib/page/secondary_page/song/model/song_url/datum.dart
@@ -17,7 +17,7 @@ class Datum {
   int? expi;
   String? type;
   double? gain;
-  int? peak;
+  double? peak;
   int? fee;
   dynamic uf;
   int? payed;
@@ -83,7 +83,7 @@ class Datum {
     int? expi,
     String? type,
     double? gain,
-    int? peak,
+    double? peak,
     int? fee,
     dynamic uf,
     int? payed,
